feat(coveralls): Improve instructions for activating Coveralls
This tweak ensures that an automatic run of coverage happens with the travis runs.  I also put in some better instructions for how to set up the repo on Coveralls, including a link for how to change the coverage degrease failure threshold (we know the default is quite annoying).

I hope it's also ok that my auto-PEP8 plugin cleared out a few of the extra unnecessary spaces in the files.