fix(cookiecutter): don't render layout.html
Cookiecutter was failing because the Jinja template of layout.html was conflicting with the
variables of cookiecutter. Cookiecutter uses Jinja templating to render the project variables into
template files. Because layout.html also used Jinja it was confusing cookiecutter with tons of
variables it did not have knowledge of. Long story short, if you add more jinja files that you don't
want to be hydrated with variables you can add it to the `_copy_without_render` list in
`cookiecutter.json`.

#14